==> project/src/4_initial_model.ipynb <==
import pandas as pd
import numpy as np
import warnings
from datetime import datetime
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error
from sklearn.metrics import make_scorer, mean_squared_error

warnings.filterwarnings("ignore")

# Load training data
train_df = pd.read_csv('../../generated_datasets/train_df.csv')

# Limit training set to before 2023
X_train = train_df[train_df['DATE']<'2023-01-01']
y_train = train_df[train_df['DATE']<'2023-01-01']['HOME PLUS MINUS']

X_test = train_df[train_df['DATE']>='2023-01-01']
y_test = train_df[train_df['DATE']>='2023-01-01']['HOME PLUS MINUS']

# Drop unneccesary features and the target variable (HOME PLUS MINUS)
X_train = X_train.drop(['HOME PLUS MINUS', 'SEASON', 'DATE', 'HOME', 'AWAY', 'HOME_PRIOR_GAME_DATE', 'AWAY_PRIOR_GAME_DATE', 'HOME TEAM GP', 'AWAY TEAM GP'], axis=1)
X_test = X_test.drop(['HOME PLUS MINUS', 'SEASON', 'DATE', 'HOME', 'AWAY', 'HOME_PRIOR_GAME_DATE', 'AWAY_PRIOR_GAME_DATE', 'HOME TEAM GP', 'AWAY TEAM GP'], axis=1)

X_train.head()

# Create a Random Forest Regressor object
rf = RandomForestRegressor(n_estimators=100, random_state=42)

# Train the model using the training sets
rf.fit(X_train, y_train)

# Use the model to make predictions
y_pred = rf.predict(X_test)

# Evaluate the model
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error: {mse}")

# Create a Random Forest Regressor object
rf = RandomForestRegressor(random_state=42)

# Define the parameter grid
param_grid = {
    'n_estimators': [100],
    'max_features': ['auto', 'sqrt', 'log2'],
    'max_depth': [None, 5, 10],
    # 'min_samples_split': [2, 5, 10],
    # 'min_samples_leaf': [1, 2, 4],
    'bootstrap': [True, False]
}

# Define a scorer
scorer = make_scorer(mean_squared_error, greater_is_better=False)

# Define the grid search
grid_search = GridSearchCV(estimator=rf, param_grid=param_grid, cv=5, scoring=scorer, n_jobs=-1)

# Fit the grid search
grid_search.fit(X_train, y_train)

# Get the best parameters
best_params = grid_search.best_params_

print(f"Best parameters: {best_params}")

# Fit the model with the best parameters
rf_best = RandomForestRegressor(**best_params, random_state=42)
rf_best.fit(X_train, y_train)

# Get the best parameters
best_params = grid_search.best_params_

# Set the number of estimators higher 
best_params['n_estimators'] = 300

print(f"Best parameters: {best_params}")

# Fit the model with the best parameters
rf_best = RandomForestRegressor(**best_params, random_state=42)
rf_best.fit(X_train, y_train)

# Use the model to make predictions
y_pred = rf_best.predict(X_test)

# Evaluate the model
mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error: {mse}")

# Establishing a baseline: comparing our initial model against simply using the home spread value (taken from the betting sites) 
mean_squared_error(list(X_test['HOME SPREAD']), -y_test)

154 < 159. This outcome suggests that simply using the betting line generates slightly better predictions than our model, which uses the betting line along with additional custom features. Let's now see how important the various input features were in making our decision trees.

# Get importance
importances = rf_best.feature_importances_

# Summarize feature importances

# To sort and visualize the features by their importance
feature_importances = pd.DataFrame(rf_best.feature_importances_,
                                   index = X_train.columns,
                                    columns=['importance']).sort_values('importance', ascending=False)

feature_importances

Unsurprisingly, the HOME SPREAD feature (from the betting sites) is by far the most important in our model's predictions. We need to find a way to beat the spread performance, so future models will not incorporate this feature!
